refactor: move function to utils
